moved the classes to separate files and added some class documentation
